Update full_prep.py change filter for empty summary
Orig filters empty summaries (no abstract sentences assigned to section based on rouge) out. This info can not be known when using the method on test data, because reference summary would have to be available. Changed it so the filter is only applied to train and val set, but not test set